changes to clang completion and verticle line
deoplete-clang:
added stronger warning flags

supertab:
the default behavior of supertab is that tab goes backwards through the
list. They specifically pointed this out in the README and gave a
reasonable-sounding justification for it, so I gave it a chance. That
chance is over.

user:
modified colorcolumn so that the highlight only appears if I'm about to
go over the character count